GitHub_with_GCP/Lab2/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
scikit-learn==1.5.2
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2
python-dotenv==1.0.1
pytest==8.3.3
flake8==7.1.1
google-cloud-storage
python-dotenv